Update build system
The script files on the design system were being processed by the dependant's browserify (because of the definition on package.json).
When the plugins and presets between dependant/dependency were different the build would throw a module not found error.
This improved build system changes the browserify transform key so that the definitions are not picked up and also includes the transpiled files to prevent errors when the dependant's transpiler is different.